examples/esp_zigbee_ota/ota_client/managed_components/espressif__esp_delta_ota/detools/requirements.txt
humanfriendly
bitstruct
pyelftools
zstandard
lz4
heatshrink2
nala
